+++ src/main/java/com/vsu/Repository/FilmRepository.java
@@ -3,6 +3,6 @@ package com.vsu.Repository;
 import com.vsu.Models.Film;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FilmRepository extends JpaRepository<Film,Integer> {
+public interface FilmRepository extends JpaRepository<Film,Long> {
 
 }
